Test Formatting Required on index.html
# Conflicts:
#	index.html

There are a few issues here that differ from the mockup.

The name and date on the mockup don't appear as headings (bigger and
always bolded) but as regular paragraphs. For the text on these two
lines to match the mockup, the text on each line needs to be within a
<p> tag.
Also, the word "By" shouldn't be bolded. Fixing this means moving the
<b> tag to after "By" so that only the name is bolded.
The date shouldn't be bolded, so the <b> and </b> tags aren't needed.